Split provenance formatting from tracing imports
