+++ features/llm/clients/auto.client.ts
@@ -25,7 +25,7 @@
  */
 import { SysEnv } from '@app/env';
 import { Oops } from '@app/nest/exceptions/oops';
-import { mergeProvenanceLlmTags } from '@app/nest/trace/provenance-tags';
+import { mergeProvenanceLlmTags } from '@app/nest/trace/provenance-context';
 
 import { getModel } from '../types/model.types';
 import { google, openrouter, vertex, vertexGlobal } from './llm.clients';
